requirements.txt
aiofiles==24.1.0
aioredis==2.0.1
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
babel==2.17.0
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
ecdsa==0.19.1
et_xmlfile==2.0.0
Faker==37.1.0
fastapi==0.115.12
fastapi-admin==1.0.4
h11==0.16.0
httptools==0.6.4
idna==3.10
iso8601==2.1.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
passlib==1.7.4
peewee==3.18.1
pendulum==3.1.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
pypika-tortoise==0.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
tortoise-orm==0.25.0
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
